+++ src/types/index.ts
@@ -16,6 +16,7 @@ export type Theme =
     | 'dark'
     | 'light'
     | 'monokai-dark'
+    | 'visual-studio-dark'
     | 'visual-studio-light'
     | 'atom-dark'
     | 'atom-light'